Renamed old and new convolution implementations
