+++ src/operator/convolution.cu
@@ -1,8 +1,8 @@
 /*!
- * Copyright (c) 2015 by Contributors
+ * Copyright (c) 2017 by Contributors
  * \file convolution.cu
  * \brief
- * \author Bing Xu
+ * \author Bing Xu, Jun Wu
 */
 
 #include "./convolution-inl.h"
@@ -10,9 +10,60 @@
 #if MXNET_USE_CUDNN == 1
 #include "./cudnn_convolution-inl.h"
 #endif  // MXNET_USE_CUDNN
+#include "./nn/im2col.h"
 
 namespace mxnet {
 namespace op {
+
+template<typename xpu, typename DType>
+inline
+void ConvolutionOp<xpu, DType>::ConvIm2Col(const DType* data_ptr,
+                                           const TShape& data_shape,
+                                           DType* col_buffer_ptr,
+                                           const TShape& col_buffer_shape,
+                                           const gpu& dev_gpu) const {
+  if (!force_nd_im2col_ && num_spatial_axes_ == 2) {
+    im2col_gpu(data_ptr, conv_in_channels_, data_shape[2], data_shape[3],
+               param_.kernel[0], param_.kernel[1], param_.pad[0], param_.pad[1],
+               param_.stride[0], param_.stride[1], param_.dilate[0], param_.dilate[1],
+               col_buffer_ptr);
+  } else {
+    im2col_nd_gpu(data_ptr, num_spatial_axes_, num_kernels_im2col_,
+                  reinterpret_cast<const int*>(&(data_shape[1])),
+                  reinterpret_cast<const int*>(col_buffer_shape.data()),
+                  reinterpret_cast<const int*>(param_.kernel.data()),
+                  reinterpret_cast<const int*>(param_.pad.data()),
+                  reinterpret_cast<const int*>(param_.stride.data()),
+                  reinterpret_cast<const int*>(param_.dilate.data()),
+                  col_buffer_ptr);
+  }
+}
+
+template<typename xpu, typename DType>
+inline
+void ConvolutionOp<xpu, DType>::ConvCol2Im(const DType* col_buffer_ptr,
+                                           const TShape& col_buffer_shape,
+                                           DType* data_ptr,
+                                           const TShape& data_shape,
+                                           OpReqType req,
+                                           const gpu& dev_gpu) const {
+  if (!force_nd_im2col_ && num_spatial_axes_ == 2) {
+    col2im_gpu(col_buffer_ptr, conv_in_channels_, data_shape[2], data_shape[3],
+               param_.kernel[0], param_.kernel[1], param_.pad[0], param_.pad[1],
+               param_.stride[0], param_.stride[1], param_.dilate[0], param_.dilate[1],
+               data_ptr, req);
+  } else {
+    col2im_nd_gpu(col_buffer_ptr, num_spatial_axes_, num_kernels_im2col_,
+                  reinterpret_cast<const int*>(&(data_shape[1])),  // skip batch dim
+                  reinterpret_cast<const int*>(col_buffer_shape.data()),
+                  reinterpret_cast<const int*>(param_.kernel.data()),
+                  reinterpret_cast<const int*>(param_.pad.data()),
+                  reinterpret_cast<const int*>(param_.stride.data()),
+                  reinterpret_cast<const int*>(param_.dilate.data()),
+                  data_ptr, req);
+  }
+}
+
 template<>
 Operator* CreateOp<gpu>(ConvolutionParam param, int dtype,
                         std::vector<TShape> *in_shape,
