+++ src/operator/convolution_v1.cu
@@ -1,0 +1,44 @@
+/*!
+ * Copyright (c) 2015 by Contributors
+ * \file convolution_v1.cu
+ * \brief
+ * \author Bing Xu
+*/
+
+#include "./convolution_v1-inl.h"
+#include <vector>
+#if MXNET_USE_CUDNN == 1
+#include "./cudnn_convolution-inl.h"
+#endif  // MXNET_USE_CUDNN
+
+namespace mxnet {
+namespace op {
+template<>
+Operator* CreateOp<gpu>(ConvolutionV1Param param, int dtype,
+                        std::vector<TShape> *in_shape,
+                        std::vector<TShape> *out_shape,
+                        Context ctx) {
+  Operator *op = NULL;
+#if MXNET_USE_CUDNN == 1
+#if 0
+  if (param.dilate.Size() == 1 && !param.cudnn_off) {
+    MSHADOW_REAL_TYPE_SWITCH(dtype, DType, {
+      op = new CuDNNConvolutionOp<DType>(param, *in_shape, *out_shape, ctx);
+    })
+  } else {
+    MSHADOW_REAL_TYPE_SWITCH(dtype, DType, {
+      op = new ConvolutionV1Op<gpu, DType>(param);
+    })
+  }
+#endif
+#else
+  MSHADOW_REAL_TYPE_SWITCH(dtype, DType, {
+    op = new ConvolutionV1Op<gpu, DType>(param);
+  })
+#endif  // MXNET_USE_CUDNN
+  return op;
+}
+
+}  // namespace op
+}  // namespace mxnet
+
